requirements.txt
Django==4.2.7
django-environ==0.10.0
django-crispy-forms==2.0
crispy-tailwind==0.5.0
Pillow==10.1.0
django-bleach==3.0.1
django-compressor==4.4
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.5.0
django-debug-toolbar==4.2.0
django-summernote==0.8.20
mysqlclient==2.2.7
requests==2.31.0
django-colorfield==0.10.1 